Refactor Code out into separate files
